feat: allowlist for version numbers in YAML comments (#172)
* feat: allowlist for version numbers in YAML comments

If an environment variable `NODE_RELEASED_VERSIONS` is detected, it is
used to validate the version numbers in the YAML comments. At the
notable exception of version numbers 0.0.x and 0.1.x,  and the
`REPLACEME` placeholder, any value not in the list is reported as
invalid.

The idea is to use this in a GitHub Action. It will generally not affect
running the linter locally.